refactoring

+++ src/auth/schemas.py
@@ -1,10 +1,12 @@
+from decimal import Decimal
+
 from fastapi_users import schemas
 
 
 class UserRead(schemas.BaseUser[int]):
     username: str
     role_id: int
-    balance: float
+    balance: Decimal
 
 
 class UserCreate(schemas.BaseUserCreate):
