+++ src/balance/service.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from sqlalchemy.ext.asyncio import AsyncSession
 from fastapi import Depends
 from database.models import User
@@ -8,7 +10,7 @@ class BalanceService:
     def __init__(self, session: AsyncSession = Depends(get_async_session)):
         self.session = session
 
-    async def top_up_balance(self, user: User, amount: float) -> User:
+    async def top_up_balance(self, user: User, amount: Decimal) -> User:
         user.balance += amount
         await self.session.merge(user)
         await self.session.commit()
